Removed templatization from HttpRequest
Parsing the body is now done after the request within the HttpResponse
