--- Http/Editor.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: a9a925362893b49d4bcebd1c9446e136
-folderAsset: yes
-timeCreated: 1510915414
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
