--- Http/Editor/CreateHttpConfigMenu.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 869a5c2c8082a402b885dc4db809ba9f
-timeCreated: 1510915434
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
