--- Http/HttpError.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 8728584f8802148e79144952be26d345
-timeCreated: 1510914986
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
