+++ Http/HttpRequest.cs
@@ -3,33 +3,6 @@ using UnityEngine.Networking;
 
 namespace DUCK.Http
 {
-	public class HttpRequest<TSuccess, TError> where TSuccess : class where TError : class
-	{
-		public UnityWebRequest UnityWebRequest { get; private set; }
-		public MarkUpType MarkUpType { get; private set; }
-
-		public HttpRequest(UnityWebRequest unityWebRequest, MarkUpType markUpType)
-		{
-			UnityWebRequest = unityWebRequest;
-			MarkUpType = markUpType;
-		}
-
-		public void SetHeader(string name, string value)
-		{
-			UnityWebRequest.SetRequestHeader(name, value);
-		}
-
-		public void Send(Action<HttpResponse, TSuccess> onSuccess = null, Action<HttpResponse, TError> onError = null)
-		{
-			Http.Instance.Send(this, onSuccess, onError);
-		}
-
-		public void Send(Action<TSuccess> onSuccess = null, Action<TError> onError = null)
-		{
-			Http.Instance.Send(this, onSuccess, onError);
-		}
-	}
-
 	public class HttpRequest
 	{
 		public UnityWebRequest UnityWebRequest { get; private set; }
@@ -37,6 +10,11 @@ namespace DUCK.Http
 		public HttpRequest(UnityWebRequest unityWebRequest)
 		{
 			UnityWebRequest = unityWebRequest;
+		}
+
+		public void SetHeader(string name, string value)
+		{
+			UnityWebRequest.SetRequestHeader(name, value);
 		}
 
 		public void Send(Action<HttpResponse> onSuccess = null, Action<HttpResponse> onError = null)
